refactor: FTC-0 rename timer camera logic
