+++ lib/src/logic/timer_camera/timer_camera_state_provider.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import 'timer_camera.dart';
+
+final timerCameraStateProvider = StateNotifierProvider.autoDispose<TimerCameraStateNotifier, TimerCameraState>((ref) {
+  return TimerCameraStateNotifier();
+});
